+++ DS_2020_GUI/DS_2020_GUI/DS_console/BPIO.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "pch.h"
 namespace BPIO {
-	int db_read(FILE* , off_t , void* , size_t , size_t );
-	int db_write(FILE* , off_t , const void* , size_t, size_t );
+	size_t db_read(FILE* , off_t , void* , size_t , size_t );
+	size_t db_write(FILE* , off_t , const void* , size_t, size_t );
 };
